requirements.txt
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.3
enum34==1.1.6
Flask==0.12.2
future==0.16.0
gspread==0.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.5.0
MarkupSafe==1.0
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pymongo==3.5.1
pyOpenSSL==17.4.0
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
